remove deeplabcut.compat in favor of deeplabcut.api
