+++ deeplabcut/generate_training_dataset/multiple_individuals_trainingsetmanipulation.py
@@ -19,10 +19,9 @@ from pathlib import Path
 import numpy as np
 from tqdm import tqdm
 
-import deeplabcut.compat as compat
 import deeplabcut.generate_training_dataset.metadata as metadata
 from deeplabcut.core.config import ProjectConfig
-from deeplabcut.core.engine import Engine
+from deeplabcut.core.engine import Engine, get_project_engine
 from deeplabcut.core.weight_init import WeightInitialization
 from deeplabcut.utils import (
     auxfun_models,
@@ -300,7 +299,7 @@ def create_multianimaltraining_dataset(
 
     # load the engine to use to create the shuffle
     if engine is None:
-        engine = compat.get_project_engine(cfg)
+        engine = get_project_engine(cfg)
 
     if not (any(net in net_type for net in ("resnet", "eff", "dlc", "mob")) or engine == Engine.PYTORCH):
         raise ValueError(f"Unsupported network {net_type} for engine {engine}.")
